refactor(context): merge non-strict logic back into FakeContentGenerator
