+++ packages/core/src/core/contentGenerator.ts
@@ -23,7 +23,6 @@ import type { UserTierId, GeminiUserTier } from '../code_assist/types.js';
 import { LoggingContentGenerator } from './loggingContentGenerator.js';
 import { InstallationManager } from '../utils/installationManager.js';
 import { FakeContentGenerator } from './fakeContentGenerator.js';
-import { NonStrictFakeContentGenerator } from './nonStrictFakeContentGenerator.js';
 import { parseCustomHeaders } from '../utils/customHeaderUtils.js';
 import { determineSurface } from '../utils/surface.js';
 import { RecordingContentGenerator } from './recordingContentGenerator.js';
@@ -195,8 +194,9 @@ export async function createContentGenerator(
 ): Promise<ContentGenerator> {
   const generator = await (async () => {
     if (gcConfig.fakeResponsesNonStrict) {
-      const fakeGenerator = await NonStrictFakeContentGenerator.fromFile(
+      const fakeGenerator = await FakeContentGenerator.fromFile(
         gcConfig.fakeResponsesNonStrict,
+        { nonStrict: true },
       );
       return new LoggingContentGenerator(fakeGenerator, gcConfig);
     }
